refactor(server): split e2e tests into domain files
Split 1330-line e2e.rs into 6 focused files:
- main.rs: shared helpers (spawn_server, connect_to_play, etc.)
- status.rs: server list ping (1 test)
- login.rs: login flow, teleport confirm (2 tests)
- chat.rs: chat, commands (6 tests)
- blocks.rs: dig, place, drops, chests (8 tests)
- multiplayer.rs: multi-player sync (4 tests)
